+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Next Version
 
 New features
-- `DELETE /api/v1/products/{id}` added. (`Client::deleteProduct($id)`) ([PR #18](https://github.com/billbeeio/billbee-php-sdk/pull/18))
+- `DELETE /api/v1/products/{id}` added. (`Client::deleteProduct($id)`) ([PR #31](https://github.com/billbeeio/billbee-php-sdk/pull/31))
 
 ## 1.7.0 (15 Nov 2019)
 New features: